preprocess fix int

+++ package/chemicalchecker/core/preprocess.py
@@ -327,12 +327,12 @@ class Preprocess():
                 hf.create_dataset("keys", data=np.array(
                     keys, DataSignature.string_dtype()))
                 hf.create_dataset(
-                    "X", (len(keys), len(orderwords)), dtype=int )
+                    "X", (len(keys), len(orderwords)), dtype='int8' )
                 hf.create_dataset("features", data=np.array(
                     orderwords, DataSignature.string_dtype()))
 
             # write data in H5
-            raws = np.zeros((chunk, len(orderwords)), dtype=int )
+            raws = np.zeros((chunk, len(orderwords)), dtype='int8' )
             wordspos = {k: v for v, k in enumerate(orderwords)}
             index = 0
             for i, k in enumerate(keys):
